The Modules class

The Modules object stores the TF modules extracted from the inferred network..

The Network object stores the inferred network itself, information about the fitting process as well as graph representations of the network.

Slots

meta

A dataframe with meta data about the modules.

features

A named list with a set of fetures (genes/peaks) for each module.

params

A named list with module selection parameters.

fit

A dataframe with goodness of fit measures.

coefs

A dataframe with the fitted coefficients.

modules

A list TF modules.

graphs

Graphical representations of the inferred network.

params

A named list with GRN inference parameters.